Rename Class in camel Case to match PEP8
